Rename api user to users
